pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiops-skills"
version = "0.1.0"
description = "A marketplace of skills for AI-powered operations tasks"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Red Hat Emerging Technologies"},
]
keywords = ["aiops", "claude", "skills", "operations", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "mlflow[genai]>=3.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.0.0",
    "jsonschema>=4.0.0",
    "bandit[toml]>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/redhat-et/aiops-skills"
Repository = "https://github.com/redhat-et/aiops-skills"
Issues = "https://github.com/redhat-et/aiops-skills/issues"

[tool.hatch.build.targets.wheel]
packages = ["skills/root-cause-analysis/scripts"]

[tool.pytest.ini_options]
testpaths = ["skills/root-cause-analysis/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["scripts"]

[tool.bandit]
exclude_dirs = ["tests", ".venv", "experiments"]
skips = []
level = "medium"
confidence = "medium"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
